Update metadata extraction to support v2 project, combined into one script, remove old scripts
