--- get_new_records/README.md
@@ -1,2 +1,0 @@
-# automate_new_records
-Tool for automating the creation of new ROR records
