--- get_new_records/requirements.txt
@@ -1,14 +1,0 @@
-certifi==2024.2.2
-cffi==1.16.0
-charset-normalizer==3.3.2
-cryptography==42.0.5
-Deprecated==1.2.14
-idna==3.6
-pycparser==2.21
-PyGithub==2.2.0
-PyJWT==2.8.0
-PyNaCl==1.5.0
-requests==2.31.0
-typing_extensions==4.10.0
-urllib3==2.2.1
-wrapt==1.16.0
