+++ get_record_metadata/requirements.txt
@@ -1,0 +1,14 @@
+certifi==2024.8.30
+cffi==1.17.1
+charset-normalizer==3.3.2
+cryptography==43.0.1
+Deprecated==1.2.14
+idna==3.8
+pycparser==2.22
+PyGithub==2.4.0
+PyJWT==2.9.0
+PyNaCl==1.5.0
+requests==2.32.3
+typing_extensions==4.12.2
+urllib3==2.2.2
+wrapt==1.16.0
